lib/resolve: when forwarding, prefer to send CD=0 upstream
FIXME: in-code comments with explanation, NEWS entry, more testing.

The diff is mainly refactoring.  The important change is for forwarding,
use CD from client instead of always CD=1.  I believe this is a bit
better for us.  There are strategies that seem even better but way more
complex (re-sending with different CD in some situations).